Lombok plugin and Data annotation were added + all ids refactored to type Long

+++ src/main/java/com/cookCycle/repository/FavoriteRepository.java
@@ -3,6 +3,6 @@ package com.cookCycle.repository;
 import com.cookCycle.model.Favorite;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FavoriteRepository extends CrudRepository<Favorite, Integer>{
+public interface FavoriteRepository extends CrudRepository<Favorite, Long>{
 
 }
