+++ src/main/java/com/cookCycle/repository/IngredientsInRecipesRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface IngredientsInRecipesRepository extends CrudRepository<IngredientsInRecipes, Integer>{
+public interface IngredientsInRecipesRepository extends CrudRepository<IngredientsInRecipes, Long>{
 
 }
