+++ src/main/java/com/cookCycle/service/IIngredientsInRecipesService.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 public interface IIngredientsInRecipesService {
     List<IngredientsInRecipes> getAllIngredientsInRecipes();
-    IngredientsInRecipes getIngredientsInRecipesById(int ingredientsInRecipesId);
+    IngredientsInRecipes getIngredientsInRecipesById(Long ingredientsInRecipesId);
 
 }
